refactor: add overaly to registration and signin
